% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project4.R
\name{align}
\alias{align}
\title{Title Show the alignment of a mapped read}
\usage{
align(ref, bam, cigar)
}
\arguments{
\item{ref}{The genome}

\item{bam}{My BAM/SAM files}

\item{cigar}{The alignment that I want}
}
\value{
The position of the alignment
}
\description{
Title Show the alignment of a mapped read
}
\examples{
align(('AAGTCGTAGAATACGTAAGTCGTAGAATACGT'),('AACACGTAG..TAGCTAACACGTAG..TAGCT'),('2S5M2D2M'))
}
\author{
Michela Francesconi\cr Politecnico di Milano\cr Maintainer: Michela
Francesconi\cr E-Mail: \href{mailto:michelafrancesconi8@gmail.com}{michelafrancesconi8@gmail.com}
}
